requirements.txt
python-nmap
pymetasploit3
